requirements.txt
flask
python-dotenv
Pillow
google-generativeai
